+++ examples/multiple-dbs/main.tf
@@ -8,120 +8,28 @@ data "aws_region" "current" {}
 #####
 # VPC and subnets
 #####
-module "vpc" {
-  source  = "terraform-aws-modules/vpc/aws"
-  version = "~> 2.62"
+data "aws_vpc" "default" {
+  default = true
+}
 
-  name = "simple-vpc"
+data "aws_subnet_ids" "all" {
+  vpc_id = data.aws_vpc.default.id
+}
 
-  cidr = "10.0.0.0/16"
-
-  azs             = ["eu-west-1a", "eu-west-1b", "eu-west-1c"]
-  private_subnets = ["10.0.1.0/24", "10.0.2.0/24", "10.0.3.0/24"]
-  public_subnets  = ["10.0.101.0/24", "10.0.102.0/24", "10.0.103.0/24"]
-
-  enable_nat_gateway = false
-
-  tags = {
-    Environment = "test"
-  }
+data "aws_subnet" "public" {
+  for_each = data.aws_subnet_ids.all.ids
+  id       = each.value
 }
 
 #############
 # KMS key
 #############
-module "kms-rds" {
-  source  = "umotif-public/kms/aws"
-  version = "~> 1.0"
-
-  alias_name              = "rds-kms-test-key"
-  deletion_window_in_days = 7
-  enable_key_rotation     = true
-  policy = jsonencode(
-    {
-      "Version" : "2012-10-17",
-      "Statement" : [
-        {
-          "Sid" : "Enable IAM User Permissions",
-          "Effect" : "Allow",
-          "Principal" : {
-            "AWS" : [
-              "arn:aws:iam::${data.aws_caller_identity.current.account_id}:root",
-              data.aws_caller_identity.current.arn
-            ]
-          },
-          "Action" : "kms:*",
-          "Resource" : "*"
-        },
-        {
-          "Sid" : "Allow use of the key",
-          "Effect" : "Allow",
-          "Principal" : {
-            "Service" : ["rds.amazonaws.com", "monitoring.rds.amazonaws.com"]
-          },
-          "Action" : [
-            "kms:Encrypt",
-            "kms:Decrypt",
-            "kms:ReEncrypt*",
-            "kms:GenerateDataKey*",
-            "kms:DescribeKey"
-          ],
-          "Resource" : "*"
-        }
-      ]
-    }
-  )
-
-  tags = {
-    Environment = "test"
-  }
+data "aws_kms_key" "backup" {
+  key_id = "alias/aws/backup"
 }
 
-module "kms-backup" {
-  source  = "umotif-public/kms/aws"
-  version = "~> 1.0"
-
-  alias_name              = "backup-kms-test-key"
-  deletion_window_in_days = 7
-  enable_key_rotation     = true
-  policy = jsonencode(
-    {
-      "Version" : "2012-10-17",
-      "Statement" : [
-        {
-          "Sid" : "Enable IAM User Permissions",
-          "Effect" : "Allow",
-          "Principal" : {
-            "AWS" : [
-              "arn:aws:iam::${data.aws_caller_identity.current.account_id}:root",
-              data.aws_caller_identity.current.arn
-            ]
-          },
-          "Action" : "kms:*",
-          "Resource" : "*"
-        },
-        {
-          "Sid" : "Allow use of the key",
-          "Effect" : "Allow",
-          "Principal" : {
-            "Service" : ["backup.amazonaws.com"]
-          },
-          "Action" : [
-            "kms:Encrypt",
-            "kms:Decrypt",
-            "kms:ReEncrypt*",
-            "kms:GenerateDataKey*",
-            "kms:DescribeKey"
-          ],
-          "Resource" : "*"
-        }
-      ]
-    }
-  )
-
-  tags = {
-    Environment = "test"
-  }
+data "aws_kms_key" "rds" {
+  key_id = "alias/aws/rds"
 }
 
 #############
@@ -130,18 +38,18 @@ module "kms-backup" {
 module "aurora-mysql" {
   source = "umotif-public/rds-aurora/aws"
 
-  name_prefix   = "example-aurora-mysql"
-  database_name = "databaseName"
+  name_prefix   = "${var.name_prefix}-aurora-mysql"
+  database_name = "${var.name_prefix}mysqldb"
   engine        = "aurora-mysql"
 
-  vpc_id  = module.vpc.vpc_id
-  subnets = module.vpc.public_subnets
+  vpc_id  = data.aws_vpc.default.id
+  subnets = data.aws_subnet_ids.all.ids
 
-  kms_key_id = module.kms-rds.key_arn
+  kms_key_id = data.aws_kms_key.rds.arn
 
   instance_type = "db.t3.medium"
 
-  allowed_cidr_blocks = ["10.10.0.0/24", "10.20.0.0/24", "10.30.0.0/24"]
+  allowed_cidr_blocks = [for s in data.aws_subnet.public : s.cidr_block]
 
   tags = {
     Environment = "test"
@@ -151,20 +59,20 @@ module "aurora-mysql" {
 module "aurora-postgresql" {
   source = "umotif-public/rds-aurora/aws"
 
-  name_prefix             = "example-aurora-postgresql"
-  database_name           = "databaseName2"
+  name_prefix             = "${var.name_prefix}-postgresql"
+  database_name           = "${var.name_prefix}postgresqldb"
   engine                  = "aurora-postgresql"
   engine_version          = "11.8"
   engine_parameter_family = "aurora-postgresql11"
 
-  vpc_id  = module.vpc.vpc_id
-  subnets = module.vpc.public_subnets
+  vpc_id  = data.aws_vpc.default.id
+  subnets = data.aws_subnet_ids.all.ids
 
-  kms_key_id = module.kms-rds.key_arn
+  kms_key_id = data.aws_kms_key.rds.arn
 
   instance_type = "db.t3.medium"
 
-  allowed_cidr_blocks = ["10.10.0.0/24", "10.20.0.0/24", "10.30.0.0/24"]
+  allowed_cidr_blocks = [for s in data.aws_subnet.public : s.cidr_block]
 
   tags = {
     Environment = "test"
@@ -178,20 +86,19 @@ module "backup" {
   source = "../.."
 
   # Create a Vault
-  vault_name        = "test-rds-aurora"
-  vault_kms_key_arn = module.kms-backup.key_arn
+  vault_name        = "${var.name_prefix}-rds-aurora"
+  vault_kms_key_arn = data.aws_kms_key.backup.arn
 
   tags = {
     Environment = "test"
   }
 
   # Create a backup plan
-  plan_name = "test-backup-plan"
+  plan_name = "${var.name_prefix}-backup-plan"
 
   rules = [
     {
-      name              = "test-backup-rule"
-      target_vault_name = "test-rds-aurora"
+      name              = "${var.name_prefix}-backup-rule"
       schedule          = "cron(0 12 * * ? *)"
       start_window      = "65"
       completion_window = "190"
@@ -207,7 +114,7 @@ module "backup" {
     }
   ]
 
-  selection_name = "test-backup-selection"
+  selection_name = "${var.name_prefix}-backup-selection"
   selection_resources = [
     module.aurora-mysql.rds_cluster_arn,
     module.aurora-postgresql.rds_cluster_arn