+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="unreleased"></a>
+## [Unreleased]
+
+- Documentation updates
+- Use default KMS keys
+- Introduce Terratest
+
+
+<a name="1.1.0"></a>
+## [1.1.0] - 2020-11-18
+
+- Allow for multiple selection conditions for a backup plan ([#4](https://github.com/umotif-public/terraform-aws-backup/issues/4))
+- Update example in README ([#3](https://github.com/umotif-public/terraform-aws-backup/issues/3))
+
+
+<a name="1.0.1"></a>
+## [1.0.1] - 2020-11-13
+
+- target_vault_name reference a resource not a variable ([#2](https://github.com/umotif-public/terraform-aws-backup/issues/2))
+
+
 <a name="1.0.0"></a>
 ## 1.0.0 - 2020-11-09
 
@@ -9,4 +30,6 @@ All notable changes to this project will be documented in this file.
 - Initial commit
 
 
-[Unreleased]: https://github.com/umotif-public/terraform-aws-backup/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/umotif-public/terraform-aws-backup/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/umotif-public/terraform-aws-backup/compare/1.0.1...1.1.0
+[1.0.1]: https://github.com/umotif-public/terraform-aws-backup/compare/1.0.0...1.0.1